+++ src/main/java/game/models/components/Bullet.java
@@ -5,16 +5,16 @@ import java.io.Serializable;
 public class Bullet extends DynamicObject implements Serializable
 {
     private int damage;
-    private int senderId;
+    private long senderId;
 
-    public Bullet(int id, int senderId, int x, int y, int damage, int speed, double directionX, double directionY)
+    public Bullet(long id, long senderId, int x, int y, int damage, int speed, double directionX, double directionY)
     {
         super(id, x, y, 5, 5, speed, directionX, directionY);
         this.damage = damage;
         this.senderId = senderId;
     }
 
-    public int getSenderId()
+    public long getSenderId()
     {
         return senderId;
     }
