+++ src/main/java/game/models/components/DynamicObject.java
@@ -7,7 +7,7 @@ public class DynamicObject extends GameObject implements Serializable
     private int speed = 5;
     private double directionX;
     private double directionY;
-    public DynamicObject(int id, int x, int y, int width, int height, int speed, double directionX, double directionY)
+    public DynamicObject(long id, int x, int y, int width, int height, int speed, double directionX, double directionY)
     {
         super(id, x, y, width, height);
         setDirection(directionX, directionY);
