+++ src/main/java/game/models/components/StaticObject.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public class StaticObject extends GameObject implements Serializable
 {
-    public StaticObject(int id, int x, int y, int width, int height)
+    public StaticObject(long id, int x, int y, int width, int height)
     {
         super(id, x, y, width, height);
     }
