+++ src/main/java/game/models/components/Wall.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public class Wall extends StaticObject implements Serializable
 {
-    public Wall(int id, int x, int y, int width, int height)
+    public Wall(long id, int x, int y, int width, int height)
     {
         super(id, x, y, width, height);
     }
