+++ src/main/java/game/models/components/Zombie.java
@@ -8,7 +8,7 @@ public class Zombie extends DynamicObject
     private int targetX;
     private int targetY;
 
-    public Zombie(int id, int x, int y, int width, int height, int hp, int speed, int targetX, int targetY)
+    public Zombie(long id, int x, int y, int width, int height, int hp, int speed, int targetX, int targetY)
     {
         super(id, x, y, width, height, speed, 0, 0);
         this.targetX = targetX;
